[DAGE-66] fixed according to Naz

+++ rre-tools/src/rre_tools/approximate_search_evaluator/main.py
@@ -1,7 +1,7 @@
 import json
 import shutil
 import logging
-from typing import Dict, Any
+from typing import Any
 import subprocess
 import argparse
 from pathlib import Path
@@ -20,9 +20,9 @@ def _parse_args() -> argparse.Namespace:
     parser.add_argument(
         "--config",
         type=str,
-        help='Config file path to use for the application [default: "configs/approx_search_config.yaml"]',
+        help='Config file path to use for the application [default: "configs/approximate_search_evaluator/default.yaml"]',
         required=False,
-        default="configs/approx_search_config.yaml",
+        default="configs/approximate_search_evaluator/default.yaml",
     )
 
     parser.add_argument('-v', '--verbose', action='store_true',
@@ -37,12 +37,12 @@ def add_vector(rating_filename: str | Path,
     Parse the writer output and add vectors
     under the '$vector' field inside each query's placeholders.
     """
-    log.info("Loading rating file: %s", rating_filename)
+    log.debug("Loading rating file: %s", rating_filename)
     with open(Path(rating_filename), "r", encoding="utf-8") as f:
-        rating_data: Dict[str, Any] = json.load(f)
+        rating_data: dict[str, Any] = json.load(f)
 
-    log.info("Loading embeddings from: %s", embedding_filename)
-    embeddings: Dict[str, str] = {}
+    log.debug("Loading embeddings from: %s", embedding_filename)
+    embeddings: dict[str, str] = {}
     with open(Path(embedding_filename), "r", encoding="utf-8") as f:
         for line in f:
             line_json = json.loads(line)
@@ -55,19 +55,19 @@ def add_vector(rating_filename: str | Path,
             placeholders = query_dict.get("placeholders", {})
             query_text = placeholders.get("$query", "")
             query_id = datastore.query_text_to_query_id.get(query_text)
-
+            log.debug("Query_id: %s", query_id)
             if query_id and (query_id in embeddings):
                 placeholders["$vector"] = embeddings[query_id]
                 updated_queries += 1
 
             query_dict["placeholders"] = placeholders
 
-    log.info("Updated %d queries with vectors.", updated_queries)
+    log.debug("Updated %d queries with vectors.", updated_queries)
 
     with open(rating_filename, "w", encoding="utf-8") as f:
         json.dump(rating_data, f, indent=2, ensure_ascii=False)
 
-    log.info("Written updated ratings back to %s", rating_filename)
+    log.debug("Written updated ratings back to %s", rating_filename)
     return
 
 def setup_rre(eval_folder : Path, search_engine_type : str, version : str) -> None:
@@ -114,7 +114,7 @@ def main() -> None:
     conf_sets_folder = rre_resources_folder / "configuration_sets"
     ratings_folder = rre_resources_folder / "ratings"
 
-    if eval_folder.exists() and eval_folder.is_dir():
+    if eval_folder.is_dir():
         shutil.rmtree(eval_folder)
 
     setup_rre(eval_folder, config.search_engine_type, config.search_engine_version)
@@ -137,15 +137,16 @@ def main() -> None:
             json.dump(to_dump, f, indent=2, ensure_ascii=False)
 
 
-    log.info("Initializing DataStore")
+    log.debug("Initializing DataStore")
     data_store = DataStore()
 
     if config.ratings_path is not None:
-        log.info("Using the existing ratings file...")
+        log.debug("Using the existing ratings file...")
         ratings_folder.mkdir(parents=True, exist_ok=True)
         shutil.copy(config.ratings_path, ratings_folder / "ratings.json")
+        ratings_file = config.ratings_path
     else:
-        log.info("Writing initial ratings file with RreWriter...")
+        log.debug("Writing initial ratings file with RreWriter...")
         writer = RreWriter(
             writer_config=WriterConfig(
                 index=config.collection_name,
@@ -156,10 +157,11 @@ def main() -> None:
             )
         )
         writer.write(ratings_folder, data_store)
+        ratings_file = ratings_folder / "ratings.json"
 
     if config.embeddings_folder is not None:
-        log.info("Adding vectors to ratings file...")
-        add_vector(config.query_template, config.embeddings_folder / "queries_embeddings.jsonl", data_store)
+        log.debug("Adding vectors to ratings file...")
+        add_vector(ratings_file, config.embeddings_folder / "queries_embeddings.jsonl", data_store)
     else:
         log.warning("No embeddings folder was specified. If the specified templates has a '$vector' placeholder, this "
                     "will break RRE evaluation.")
@@ -171,8 +173,9 @@ def main() -> None:
     shutil.copy(eval_folder / "target" / "rre" / "evaluation.json", config.output_destination / "rre_evaluation_results.json")
     log.info(f"Evaluation file saved to `{config.output_destination}/` directory.")
 
-    shutil.rmtree(eval_folder)
-    log.debug("RRE evaluation folder cleaned up.")
+    if not args.verbose:
+        shutil.rmtree(eval_folder)
+        log.debug("RRE evaluation folder not cleaned up.")
 
 
 if __name__ == "__main__":
